setups tests, funcs, and common lib
