+++ Deluxxe.sln
@@ -1,0 +1,47 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{690AC2F1-77C0-4D25-A3E7-C4ADF4DA9FF2}"
+	ProjectSection(SolutionItems) = preProject
+		Directory.Build.props = Directory.Build.props
+		Directory.Build.targets = Directory.Build.targets
+		Directory.Packages.props = Directory.Packages.props
+		README.md = README.md
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Deluxxe", "src\Deluxxe.csproj", "{7C951B15-BCF4-4D9F-BD85-09A45933A2BF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeluxxeFunc", "src\AzureFunctions\DeluxxeFunc.csproj", "{958F3B03-3631-4DA2-BBFB-FE7502226882}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Test", "Test", "{70E5CE74-5DD1-4745-9318-D327058BF3EF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Deluxxe.Tests", "test\deluxxe.tests\Deluxxe.Tests.csproj", "{5F3B98AA-EDB3-4389-B113-296660DA08FE}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{958F3B03-3631-4DA2-BBFB-FE7502226882}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{958F3B03-3631-4DA2-BBFB-FE7502226882}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{958F3B03-3631-4DA2-BBFB-FE7502226882}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{958F3B03-3631-4DA2-BBFB-FE7502226882}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5F3B98AA-EDB3-4389-B113-296660DA08FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5F3B98AA-EDB3-4389-B113-296660DA08FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5F3B98AA-EDB3-4389-B113-296660DA08FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5F3B98AA-EDB3-4389-B113-296660DA08FE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7C951B15-BCF4-4D9F-BD85-09A45933A2BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7C951B15-BCF4-4D9F-BD85-09A45933A2BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7C951B15-BCF4-4D9F-BD85-09A45933A2BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7C951B15-BCF4-4D9F-BD85-09A45933A2BF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{5F3B98AA-EDB3-4389-B113-296660DA08FE} = {70E5CE74-5DD1-4745-9318-D327058BF3EF}
+	EndGlobalSection
+EndGlobal
