+++ Directory.Build.props
@@ -1,0 +1,1 @@
+<Project/>
