+++ Directory.Packages.props
@@ -1,0 +1,17 @@
+<Project>
+    <PropertyGroup>
+        <ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
+    </PropertyGroup>
+    <ItemGroup>
+<!--        <PackageVersion Include="bogus" Version="35.6.1" />-->
+        <PackageVersion Include="Microsoft.Azure.Functions.Worker" Version="2.0.0" />
+        <PackageVersion Include="Microsoft.Azure.Functions.Worker.Extensions.Http" Version="3.2.0" />
+        <PackageVersion Include="Microsoft.Azure.Functions.Worker.Sdk" Version="2.0.0" />
+
+        <!-- test dependencies-->
+        <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
+        <PackageVersion Include="xunit" Version="2.9.2" />
+        <PackageVersion Include="xunit.runner.visualstudio" Version="3.0.0" />
+        <PackageVersion Include="coverlet.collector" Version="6.0.3" />
+    </ItemGroup>
+</Project>
