+++ src/AzureFunctions/host.json
