--- src/ZipDeploy-AzFunc.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZipDeploy-AzFunc", "ZipDeploy-AzFunc.csproj", "{958F3B03-3631-4DA2-BBFB-FE7502226882}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{958F3B03-3631-4DA2-BBFB-FE7502226882}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{958F3B03-3631-4DA2-BBFB-FE7502226882}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{958F3B03-3631-4DA2-BBFB-FE7502226882}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{958F3B03-3631-4DA2-BBFB-FE7502226882}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
